% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{asymptotic_exponential}
\alias{asymptotic_exponential}
\title{Exponential change from intercept toward asymptote.}
\usage{
asymptotic_exponential(x, intercept, asymptote, beta)
}
\arguments{
\item{x}{Numeric value of time.}

\item{intercept}{Numeric value of intercept at t = 0.}

\item{asymptote}{Numeric asymptote of change.}

\item{beta}{Numeric rate of change.}
}
\value{
Numeric value fo change at t = x.
}
\description{
Exponential change from intercept toward asymptote.
}
\examples{
asymptotic_exponential(x = .3, intercept = 2, asymptote = 1, beta = 1)
}
