% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{colours5}
\alias{colours5}
\title{Saturated colours}
\usage{
colours5(colour = NULL)
}
\arguments{
\item{colour}{a string containing the desired colour. One of: "purple", "green", "blue", "yellow", or "red".
NULL returns all colours.}
}
\value{
A character vector of hex codes of length 1 if colour supplied, or length 5 if not.
}
\description{
Saturated colours
}
\examples{
data <- data.frame(a = LETTERS[1:3], b = 1:3)
plot <- ggplot(data = data, aes(x = a, fill = a, y = b)) +
geom_point()
plot
plot + scale_fill_manual(values = c(colours5(purple = "blue",
colours5("yellow"),
colours5("red")))
}
