% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{cor_test_mat}
\alias{cor_test_mat}
\title{Matrix of p-values for hypothesis testing of correlation coefficients}
\usage{
cor_test_mat(x, method = "pearson")
}
\arguments{
\item{x}{A wide dataframe with columns containing numeric variables to be tested}

\item{method}{Method of calculating correlation coefficient supplied to cor.test.
One of: "pearson", "kendall", "spearman".}
}
\value{
A symmetric numeric matrix containing p-values for the correlation coefficients calculated for
each pair of variables in x.
}
\description{
Matrix of p-values for hypothesis testing of correlation coefficients
}
\examples{
x <- data.frame(a = rnorm(n = 10), b = rnorm(n = 10), c = rnorm(n = 10))
cor_test_mat(x = x, method = "spearman")
}
