% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{double_exponential}
\alias{double_exponential}
\title{Exponential change}
\usage{
double_exponential(x, intercept, asymptote, alpha, beta)
}
\arguments{
\item{x}{Numeric value(s) of time.}

\item{intercept}{Numeric intercept at t = 0.}

\item{asymptote}{Numeric level of asymptote of change.}

\item{alpha}{Numeric rate of change.}

\item{beta}{Numeric rate of change.}
}
\value{
Numeric value of change at t = x.
}
\description{
Exponential change
}
\examples{
double exponential(x = 3, intercept = 1, asymptote = 5, alpha = 5, beta = 7)
}
