% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{enframe_prop_integer}
\alias{enframe_prop_integer}
\title{Obtain the proportion of each integer in a set of values}
\usage{
enframe_prop_integer(y)
}
\arguments{
\item{y}{Vector of observed values or array of posterior predictive distributions.}
}
\value{
Dataframe containing proportion of each integer contained in y.
}
\description{
Obtain the proportion of each integer in a set of values
}
\examples{
data <- data.frame(y = rpois(n = 1000, lambda = 3))
enframe_prop_integer(y = data$y)
}
