% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{fisher_r_to_z}
\alias{fisher_r_to_z}
\title{Significance testing of the difference of correlation coefficients to a null hypothesis of 0}
\usage{
fisher_r_to_z(r1, r2, n1, n2, return = "p")
}
\arguments{
\item{r1}{Numeric value of a correlation coefficient.}

\item{r2}{Numeric value of another correlation coefficient.}

\item{n1}{The number of samples used to calculate r1.}

\item{n2}{The number of samples used to calculate r2.}

\item{return}{A string indicating the desired output. "p" returns a numeric p-value.
"k" returns a length-2 numeric vector containing the z-scores of each correlation coefficient.
"z" returns the test statistic used in the difference between the correlation coefficients.}
}
\value{
See return parameter.
}
\description{
Significance testing of the difference of correlation coefficients to a null hypothesis of 0
}
\examples{
r1 <- runif(n = 1, min = -1, max = 1)
r2 <- runif(n = 1, min = -1, max = 1)
fisher_r_to_z(r1 = r1, r2 = r2, n1 = 10, n2 = 10, return = "p")
}
