% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{generalized_logistic}
\alias{generalized_logistic}
\title{Logistic change}
\usage{
generalized_logistic(A, K, B, Q, M, t, v)
}
\arguments{
\item{A}{The left asymptote.}

\item{K}{The right asymptote.}

\item{B}{Maximum rate of change.}

\item{Q}{Translation along the x axis (useful in specification of y at t = 0).}

\item{M}{Alternative method of translation along the x axis; starting time.}

\item{t}{Time.}

\item{v}{Asymptotic skew of change.}
}
\value{
Numeric value of y at t = t.
}
\description{
Logistic change
}
\examples{
generalized_logistic(A = 1, K = 10, B = 1, Q = 10, M = 0, t = seq(0, 10, .1), v = .5)
}
