% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(p)
}
\arguments{
\item{p}{A probability between 0 and 1 to be tranformed to unbounded parameter space.}
}
\value{
A numeric value.
}
\description{
Logit
}
\examples{
logit(p = .75)
}
