% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{nature_palette}
\alias{nature_palette}
\title{Colour palettes used in nature journals}
\usage{
nature_palette(colour = NULL, n = NULL, gradn)
}
\arguments{
\item{colour}{A string indicating the hue of the desired pallette. Must be one of:
"background", "grey", "olive", "green", "teal", "blue", "purple", "red", "orange", "yellow".
If NULL (the default), returns all hues.}

\item{n}{an integer between 1 and 6 inclusive. Indicates the index of the palette desired.}

\item{gradn}{(Optional) an integer between 1 and 6 inclusive. If supplied, will override n and
return a monochromatic palette of hue colour and length gradn.}
}
\value{
If neither colour nor n supplied, a list of legnth 10 with each element containing
a character vector of hex codes of length 6. If colour but not n supplied, a character vector of
hex codes of length 6.
}
\description{
Colour palettes used in nature journals
}
\examples{
data <- data.frame(a = LETTERS[1:3], b = 1:3)
plot <- ggplot(data = data, aes(x = a, fill = a, y = b)) +
geom_point()
plot
plot + scale_fill_manual(values = c(nature_palette(colour = "blue", n = 3),
nature_palette(colour = "purple", n = 3),
nature_palette(colour = "teal", n = 3)))
}
