% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro-package.R
\docType{package}
\name{neuro-package}
\alias{neuro}
\alias{neuro-package}
\title{neuro: functions for analyses in behavioural neuroscience}
\description{
A working toolkit, published so collaborators can install it rather than
receive it as emailed script files. See the README for what is here, what
is reliable, and what has sharp edges.
}
\author{
\strong{Maintainer}: Davin Peart \email{dpeart@uoguelph.ca}

Authors:
\itemize{
  \item Davin Peart \email{dpeart@uoguelph.ca}
}

}
\keyword{internal}
