% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{plot_cor_mat}
\alias{plot_cor_mat}
\title{Plot of correlation matrix showing coefficients and significance}
\usage{
plot_cor_mat(
  rmat,
  pmat,
  sepvars = NULL,
  show_legend = T,
  lwd = 0,
  mid_fill = "white",
  axis_rel = 0.7,
  legend_rel = 0.7,
  guide_rel = 0.9,
  fill_cols = c("#657EA1", "#FA819E"),
  line_col = "white",
  rsize = 3,
  psize = 2
)
}
\arguments{
\item{rmat}{Numeric matrix of correlation coefficients.}

\item{pmat}{Numeric matrix of identical structure to rmat containing correlation coefficient p-values.}

\item{sepvars}{Character vector of dimnames of rmat and pmat to be moved to the edge of the plot.}

\item{show_legend}{Logical indicating whether to plot a legend for colour map of correlation coefficient.}

\item{lwd}{Linewidth of squares containing correlation coefficients.}

\item{mid_fill}{String indicating colour for a correlation coefficient of 0.}

\item{axis_rel}{Relative size of axis text.}

\item{legend_rel}{Relative size of legend text.}

\item{guide_rel}{Relative size of legend title.}

\item{fill_cols}{Character vector of length 2 containing colours mapped to correlation coefficients of -1 and 1.}

\item{line_col}{Colour of the lines separating the squares containing each correlation coefficient.}

\item{rsize}{Size of the text indicating the values of the correlation coefficients.}

\item{psize}{Size of the text indicating significance.}
}
\value{
A ggplot depicting the correlation matrix
}
\description{
Plot of correlation matrix showing coefficients and significance
}
\examples{
x <- data.frame(a = rnorm(n = 10), b = rnorm(n = 10), c = rnorm(n = 10))
rmat <- cor(x = x)
pmat <- cor_test_mat(x = x, method = "spearman")
plot_cor_pmat(rmat = rmat, pmat = pmat)
}
