% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions.R
\name{plot_importance}
\alias{plot_importance}
\title{Scree plot of explained variance from prcomp object}
\usage{
plot_importance(pc, ncomp = 9, nret = 3, return = "plot")
}
\arguments{
\item{pc}{PCA computed using prcomp.}

\item{ncomp}{Number of components up to which to display.}

\item{nret}{Number of components to retain.}

\item{return}{A string, either "data" or "plot".}
}
\value{
See return.
}
\description{
Scree plot of explained variance from prcomp object
}
\examples{
data <- data.frame(a = rnorm(n = 20), b = rnorm(n = 20), c = rnorm(n = 20), d = rnorm(n = 20))
pc <- prcomp(x = data)
}
