% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of significance testing of differences between two matrices of correlation coefficients}
\usage{
plot_volcano(
  cor1,
  cor2,
  n1,
  n2,
  plab = 0.1,
  return = "graph",
  colours = c("#FA819E", "#5f83ea", "grey70")
)
}
\arguments{
\item{cor1}{Numeric matrix of correlation coefficients.}

\item{cor2}{Numieric matrix of correlation coefficients with identical dimensions to cor1.}

\item{plab}{Numeric value of -log10(p) above which points will be labeled.}

\item{return}{A string indicating whether to return a "graph" or "data".}

\item{colours}{A character vector of length 3 containing colours to fill negative, small, and positive fold-change values.}
}
\value{
See parameter return.
}
\description{
Volcano plot of significance testing of differences between two matrices of correlation coefficients
}
\examples{
cor1 <- cor(x = data.frame(a = rep(c("a", "b"), each = 10), b = rnorm(n = 20), c = rnorm(n = 20), d = rnorm(n = 20))[1:10, 2:4])
cor2 <- cor(x = data.frame(a = rep(c("a", "b"), each = 10), b = rnorm(n = 20), c = rnorm(n = 20), d = rnorm(n = 20))[11:20, 2:4])
plot_volcano(cor1 = cor1, cor2 = cor2)
}
