% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{pp_stat_dens}
\alias{pp_stat_dens}
\title{Plot density of two descriptive statistics of draws against observed values}
\usage{
pp_stat_dens(
  yrep,
  y,
  nbin = 4,
  stat1 = mean,
  stat2 = sd,
  fill = "yellow",
  y_col = "red",
  linetype = "dotted",
  adjust = 1,
  linewidth = 0.5,
  stroke = 0.5,
  ynu1 = NULL,
  ynu2 = NULL
)
}
\arguments{
\item{yrep}{Matrix of posterior predictive distributions with the same number of columns as rows in y.}

\item{y}{Vector of observed values.}

\item{nbin}{Number of bins of density cutoffs.}

\item{stat1}{Function to be performed on y and each draw of yrep.}

\item{stat2}{Function to be performed on y and each draw of yrep.}

\item{fill}{Colour of the gradient mapped to the density of descriptives. See ?nature_palette or ?springer.}

\item{y_col}{Colour of plotted values of y. See ?nature_palette or ?springer.}

\item{linetype}{Linetype of the plot of y passed to geom_segment.}

\item{adjust}{Bandwidth of the density boundaries passed to geom_density2d_filled.}

\item{linewidth}{Linewidth of the plot of y passed to geom_segment.}

\item{stroke}{Linewidth of the shape of y passed to geom_point.}

\item{ynu1}{(Optional) Vector of previously observed values.}

\item{ynu2}{(Optional) Vector of previously observed values.}
}
\value{
A ggplot.
}
\description{
Plot density of two descriptive statistics of draws against observed values
}
\examples{
data <- data.frame(y = rnorm(n = 100))
draws <- matrix(data = rnorm(n = 100 * 2000), nrow = 100)
pp_stat_dens(yrep = draws, y = data$y)
}
