% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam_functions.R
\name{rskellam}
\alias{rskellam}
\title{Random somplaing from the skellam distribution}
\usage{
rskellam(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Value of the mean.}

\item{sigma}{Value of the difference between the absolute value of the mean and variance.}
}
\value{
A numeric vecotr of samples of length n.
}
\description{
Random somplaing from the skellam distribution
}
\examples{
rskellam(n = 100, mu = 2, sigma = 1)
}
