% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam_functions.R
\name{rskellamzi}
\alias{rskellamzi}
\title{Random sampling from the zero-inflated skellam distribution}
\usage{
rskellamzi(n, mu, sigma, zeta)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Value of the mean.}

\item{sigma}{Value of the difference between the absolute value of the mean and variance.}

\item{zeta}{Probability of zero occuring.}
}
\value{
A Numeric vector of samples of length n.
}
\description{
Random sampling from the zero-inflated skellam distribution
}
\examples{
rskellamzi(n = 100, mu = 3, sigma = 1, zeta = .3)
}
