% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{springer}
\alias{springer}
\title{Colours from Springer journals}
\usage{
springer(N, gradn)
}
\arguments{
\item{N}{An integer from 1 to 8 indicating the length of the desired palette.}

\item{gradn}{(Optional) An integer from 1 to 6 indicating the desired darkness &
saturation of the palette.}
}
\value{
A character vector of length N containing hex codes of the desired palette.
}
\description{
Colours from Springer journals
}
\examples{
ggplot2::ggplot(data = data.frame(a = c("a", "b"), b = c(3, 4))) +
ggplot2::geom_col(ggplot2::aes(x = a, y = b, fill = a), colour = "white") +
ggplot2::scale_fill_manual(values = springer(N = 2)) +
theme_nature()
}
