% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_cell}
\alias{theme_cell}
\title{Theme for ggplot2 inspired by journals of the cell press}
\usage{
theme_cell(background = T, xticks = "up")
}
\arguments{
\item{background}{a logical object of length 1 indicating where to show or remove the grey background}

\item{xticks}{a string of length 1 (either "up" or "down") indicating the direction of the ticks on the x axis}
}
\value{
a ggplot2 object
}
\description{
Theme for ggplot2 inspired by journals of the cell press
}
\examples{
data <- data.frame(a = LETTERS[1:3], b = 1:3)
plot <- ggplot(data = data, aes(x = a, y = b)) +
geom_point()
plot
plot + theme_cell()
}
