% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_nature}
\alias{theme_nature}
\title{Theme for ggplot2 inspired by springer nature journals}
\usage{
theme_nature()
}
\value{
a ggplot2 object
}
\description{
Theme for ggplot2 inspired by springer nature journals
}
\examples{
data <- data.frame(a = LETTERS[1:3], b = 1:3)
plot <- ggplot(data = data, aes(x = a, y = b)) +
geom_point()
plot
plot + theme_nature()
}
