% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{pp_stat_dens}
\alias{pp_stat_dens}
\title{Plot density of two descriptive statistics of draws against observed values}
\usage{
pp_stat_dens(
  yrep,
  y,
  ynu1 = NULL,
  ynu2 = NULL,
  nbin = 4,
  stat1 = mean,
  stat2 = sd,
  fill_limits = c("#FEF8F9", "#FA819E"),
  y_col = "teal"
)
}
\arguments{
\item{yrep}{Matrix of posterior predictive distributions with the same number of columns as rows in y.}

\item{y}{Vector of observed values.}

\item{ynu1}{(Optional) Vector of previously observed values.}

\item{ynu2}{(Optional) Vector of previously observed values.}

\item{nbin}{Number of bins of density cutoffs.}

\item{stat1}{Function to be performed on y and each draw of yrep.}

\item{stat2}{Function to be performed on y and each draw of yrep.}

\item{fill_limits}{Limits of the colour gradient mapped to the density of descriptives.}

\item{y_col}{Colour of plotted values of y.}
}
\value{
A ggplot.
}
\description{
Plot density of two descriptive statistics of draws against observed values
}
\examples{
data <- data.frame(y = rnorm(n = 100))
draws <- matrix(data = rnorm(n = 100 * 2000), nrow = 100)
pp_stat_dens(yrep = draws, y = data$y)
}
