% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{ribbon_epreds}
\alias{ribbon_epreds}
\title{Visualization of the posterior of the expected value by group.}
\usage{
ribbon_epreds(
  brmsfit,
  x,
  y,
  wrap,
  grid,
  method = median,
  probs = c(0.5, 0.89),
  plot = "preds",
  points = T,
  hline = F,
  marginalize = F,
  nrow = NULL
)
}
\arguments{
\item{brmsfit}{Fit of class brmsfit.}

\item{x}{Expression of variable to be mapped to the x axis.}

\item{y}{Expression of variable to be mapped to the y axis (the dependent variable).}

\item{wrap}{Expression of variable to be passed to facet_wrap.}

\item{grid}{Two-sided formula to be passed to facet_grid. The LHS should contain the variable to be
displayed in the rows of the grid and the RHS should contain the variable to be mapped to the columns.}

\item{method}{A function indicating the method of calculating the point estimate of the posterior.}

\item{probs}{A numeric vector of length 2 indicating the length of the CIs.}

\item{plot}{A string indicating whether to plot "preds" or "reps".}

\item{points}{Logical indicating whether to plot points.}

\item{hline}{Logical indicating whether to draw a line at x = 0.}

\item{marginalize}{Logical indicating whether samples should be marginalized within draws. If there are multiple
draws within each component of the specified design (e.g., repeated measures within subjects), then marginalize should be TRUE.}

\item{nrow}{Number of rows passed to facet_wrap.}
}
\value{
A list of length 3 containing the specified plot, a dataframe containing samples bound to the original observations,
and a summarized dataframe containing point estimates and CIs for the specified design.
}
\description{
Visualization of the posterior of the expected value by group.
}
\examples{
data("iris")
sepal <- brms::brm(Sepal.Length ~ 1 + Sepal.Width + (1|Species), data = iris,
save_pars = brms::save_pars(all = TRUE), sample_prior = TRUE)
ribbon_epreds(brmsfit = sepal, x = Sepal.Width, y = Sepal.Length, wrap = Species, nrow = 1)


}
